adds cache to geometrygroup.bounds
